break out topological-sort.ts from metamodel.ts (using an agent)
